Rework pages and components to reduce CNCF tie-in.
In order to improve the ability for non-CNCF users of Radar to
make their own radar sites, this moves as much CNCF related content
to the content/ tree as possible.

After this, the expectation is that the content/ and public/ trees
need to be customised for a different entity wanting Radars, rather
than having to patch the main codebase.

Over time, this may be splittable into a separate repository so that
there's the radar codebase and the radar data for a site as separate
repositories, like Landscapes are.

Signed-off-by: Daniel Silverstone <daniel.silverstone@codethink.co.uk>